training/requirements.txt
pillow>=10.3.0
requests>=2.23.0
matplotlib>=3.3